Use the sentry symfony integration
The latest version of this library allows us to stop errors being sent
using our existing configuration setup without needing to write our own
service.
